split up emitter.rs
